--- src/main/java/faang/school/postservice/exception/GlobalExceptionHandler.java
@@ -1,60 +1,0 @@
-package faang.school.postservice.exception;
-
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-@Slf4j
-@RestControllerAdvice
-public class GlobalExceptionHandler {
-    @ExceptionHandler(PostNotFoundException.class)
-    public ResponseEntity<String> handlePostNotFoundException(PostNotFoundException e) {
-        log.warn("Post error: {}", e.getMessage());
-        return ResponseEntity.status(HttpStatus.NOT_FOUND).body(e.getMessage());
-    }
-
-    @ExceptionHandler(PostIdMismatchException.class)
-    public ResponseEntity<String> handlePostIdMismatchException(PostIdMismatchException e) {
-        log.warn("Validation post error: {}", e.getMessage());
-        return ResponseEntity.badRequest().body(e.getMessage());
-    }
-
-    @ExceptionHandler(AuthorNotFoundException.class)
-    public ResponseEntity<String> handleAuthorNotFoundException(AuthorNotFoundException e) {
-        log.warn("Author error: {}", e.getMessage());
-        return ResponseEntity.status(HttpStatus.NOT_FOUND).body(e.getMessage());
-    }
-
-    @ExceptionHandler(CommentNotFoundException.class)
-    public ResponseEntity<String> handleCommentNotFoundException(CommentNotFoundException e) {
-        log.warn("Comment error: {}", e.getMessage());
-        return ResponseEntity.status(HttpStatus.NOT_FOUND).body(e.getMessage());
-    }
-
-    @ExceptionHandler(PostDtoValidationException.class)
-    public ResponseEntity<String> handleValidationException(PostDtoValidationException e) {
-        log.warn("Post dto validation exception: {}", e.getMessage());
-        return ResponseEntity.badRequest().body(e.getMessage());
-    }
-
-    @ExceptionHandler(CommentIdMismatchException.class)
-    public ResponseEntity<String> handleCommentIdMismatchException(CommentIdMismatchException e) {
-        log.warn("Validation comment error: {}", e.getMessage());
-        return ResponseEntity.badRequest().body(e.getMessage());
-    }
-
-    @ExceptionHandler(LikeAlreadyExistException.class)
-    public ResponseEntity<String> handleLikeAlreadyExistException(LikeAlreadyExistException e) {
-        log.warn("Like conflict error: {}", e.getMessage());
-        return ResponseEntity.status(HttpStatus.CONFLICT).body(e.getMessage());
-    }
-
-    @ExceptionHandler(LikeNotFoundException.class)
-    public ResponseEntity<String> handleLikeNotFoundException(LikeNotFoundException e) {
-        log.warn("Like error: {}", e.getMessage());
-        return ResponseEntity.status(HttpStatus.NOT_FOUND).body(e.getMessage());
-    }
-}
-
