+++ src/main/java/faang/school/postservice/exception/handler/GlobalExceptionHandler.java
@@ -1,10 +1,13 @@
 package faang.school.postservice.exception.handler;
 
 import faang.school.postservice.exception.AuthorNotFoundException;
+import faang.school.postservice.exception.CommentIdMismatchException;
 import faang.school.postservice.exception.CommentNotFoundException;
 import faang.school.postservice.exception.DataValidationException;
 import faang.school.postservice.exception.FileProcessException;
 import faang.school.postservice.exception.ForbiddenException;
+import faang.school.postservice.exception.LikeAlreadyExistException;
+import faang.school.postservice.exception.LikeNotFoundException;
 import faang.school.postservice.exception.PostDtoValidationException;
 import faang.school.postservice.exception.PostIdMismatchException;
 import faang.school.postservice.exception.PostNotFoundException;
@@ -29,13 +32,19 @@ public class GlobalExceptionHandler {
 
     @ExceptionHandler(PostNotFoundException.class)
     public ResponseEntity<String> handlePostNotFoundException(PostNotFoundException e) {
-        log.warn("Find error: {}", e.getMessage());
+        log.warn("Post find error: {}", e.getMessage());
         return ResponseEntity.status(HttpStatus.NOT_FOUND).body(e.getMessage());
     }
 
     @ExceptionHandler(PostIdMismatchException.class)
     public ResponseEntity<String> handlePostIdMismatchException(PostIdMismatchException e) {
-        log.warn("Validation error: {}", e.getMessage());
+        log.warn("Post validation error: {}", e.getMessage());
+        return ResponseEntity.badRequest().body(e.getMessage());
+    }
+
+    @ExceptionHandler(PostDtoValidationException.class)
+    public ResponseEntity<String> handleValidationException(PostDtoValidationException e) {
+        log.warn("Post dto validation exception: {}", e.getMessage());
         return ResponseEntity.badRequest().body(e.getMessage());
     }
 
@@ -47,7 +56,27 @@ public class GlobalExceptionHandler {
 
     @ExceptionHandler(CommentNotFoundException.class)
     public ResponseEntity<String> handleCommentNotFoundException(CommentNotFoundException e) {
-        log.warn("Comment error: {}", e.getMessage());
+        log.warn("Comment find error: {}", e.getMessage());
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).body(e.getMessage());
+    }
+
+    @ExceptionHandler(CommentIdMismatchException.class)
+    public ResponseEntity<String> handleCommentIdMismatchException(CommentIdMismatchException e) {
+        log.warn("Validation comment error: {}", e.getMessage());
+        return ResponseEntity.badRequest().body(e.getMessage());
+
+    }
+
+    @ExceptionHandler(LikeAlreadyExistException.class)
+    public ResponseEntity<String> handleLikeAlreadyExistException(LikeAlreadyExistException e) {
+        log.warn("Like conflict error: {}", e.getMessage());
+        return ResponseEntity.status(HttpStatus.CONFLICT).body(e.getMessage());
+
+    }
+
+    @ExceptionHandler(LikeNotFoundException.class)
+    public ResponseEntity<String> handleLikeNotFoundException(LikeNotFoundException e) {
+        log.warn("Find like error: {}", e.getMessage());
         return ResponseEntity.status(HttpStatus.NOT_FOUND).body(e.getMessage());
     }
 
@@ -63,12 +92,6 @@ public class GlobalExceptionHandler {
     public ErrorResponse handleFileProcessException(FileProcessException e) {
         log.error("FileProcessException", e);
         return buildResponse(e);
-    }
-
-    @ExceptionHandler(PostDtoValidationException.class)
-    public ResponseEntity<String> handleValidationException(PostDtoValidationException e) {
-        log.warn("Post dto validation exception: {}", e.getMessage());
-        return ResponseEntity.badRequest().body(e.getMessage());
     }
 
     private ErrorResponse buildResponse(Exception e) {
